Minor change convertion issue of Invoice Qty., converted data type of this column from float to int.

+++ upgrade_mandi/Zepto_PDF.py
@@ -167,7 +167,7 @@ class Zepto_PDF:
             "Total",
             "",
             "",
-            str(df["Invoice Qty."].apply(int).sum()),
+            str(df["Invoice Qty."].apply(float).sum()),
             "",
             "",
             str(df["Amount"].apply(int).sum()),
